requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.4
cryptography==46.0.4
db-dtypes==1.5.0
donfig==0.8.1.post1
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h5netcdf==1.8.1
idna==3.11
legacy-cgi==2.6.4
lxml==6.0.2
netCDF4==1.7.4
numcodecs==0.16.5
numpy==2.4.1
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-gbq==0.33.0
platformdirs==4.5.1
pooch==1.8.2
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.2
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydap==3.5.8
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-cache==1.2.1
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.3
WebOb==1.8.9
xarray==2025.12.0
zarr==3.1.5
